Adds auto complete instrux + chains info
