+++ cmd/local-ic/main.go
@@ -3,12 +3,6 @@ package main
 import (
 	"fmt"
 	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var (
-	MakeFileInstallDirectory string
 )
 
 func main() {
@@ -16,21 +10,4 @@ func main() {
 		fmt.Fprintf(os.Stderr, "Whoops. There was an error while executing your CLI '%s'", err)
 		os.Exit(1)
 	}
-}
-
-func GetDirectory() string {
-	installDir := os.Getenv("INSTALL_DIR")
-	if installDir != "" {
-		return installDir
-	}
-
-	return MakeFileInstallDirectory
-}
-
-var rootCmd = &cobra.Command{
-	Use:   "local-ic",
-	Short: "Your local IBC interchain of nodes program",
-	Run: func(cmd *cobra.Command, args []string) {
-		cmd.Help()
-	},
 }
